requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
